project1_vercel/requirements.txt
requests==2.32.3
redis==5.2.1
psycopg2-binary==2.9.10
